csum: fixed for 16bits platforms

+++ sys/net/crosslayer/inet_csum/inet_csum.c
@@ -45,16 +45,16 @@ uint16_t inet_csum_slice(uint16_t sum, const uint8_t *buf, uint16_t len, size_t 
     }
 
     for (int i = 0; i < (len >> 1); buf += 2, i++) {
-        csum += (*buf << 8) + *(buf + 1);   /* group bytes by 16-byte words
+        csum += (uint16_t)((*buf << 8) + *(buf + 1));   /* group bytes by 16-byte words
                                              * and add them*/
     }
 
     if ((accum_len + len) & 1)      /* if accumulated length is odd */
-        csum += (*buf << 8);        /* add last byte as top half of 16-byte word */
+        csum += (uint16_t)(*buf << 8);        /* add last byte as top half of 16-byte word */
 
     while (csum >> 16) {
         uint16_t carry = csum >> 16;
-        csum = (csum & 0xffff) + carry;
+        csum = (uint16_t)(csum & 0xffff) + carry;
     }
 
     DEBUG("inet_sum: new sum = 0x%04" PRIx32 "\n", csum);
